requirements.txt
# Core dependencies
flet>=0.24.0
sqlite3-to-mysql>=2.1.7

# File processing
openpyxl>=3.1.2
python-docx>=1.1.0
PyPDF2>=3.0.1

# Audio processing
pydub>=0.25.1
SpeechRecognition>=3.10.0
pyttsx3>=2.90

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Visualization
matplotlib>=3.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0

# Packaging
pyinstaller>=6.3.0